pyproject.toml
[tool.poetry]
name = "fair-bolts"
version = "0.1.0"
description = "PyTorch-Lightning Bolts for Fair Machine Learning"
authors = ["Your Name <you@example.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
pytorch-lightning = "^1.2.0"
pytorch-lightning-bolts = "^0.3.0"
torchvision = "^0.8.2"
EthicML = "^0.2.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"
pydocstyle = "^5.1.1"
pre-commit = "^2.10.1"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
